requirements.txt
pip==25.2
joblib==1.3.2
google-cloud-storage==2.17.0
google-cloud-aiplatform==1.63.0
numpy>=1.21.0,<2.0.0
scipy>=1.12.0
pandas>=2.0.0,<3.0.0
scikit-learn==0.24